todo_app-main/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9          # ← add this (binary version, no compilation needed)
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==42.0.0
pydantic==2.5.3
boto3==1.34.0